ci: Lint check

+++ scripts/database/settings/SettingsProfile.cs
@@ -147,7 +147,7 @@ public partial class SettingsProfile
     /// Degrees to rotate the cursor by every second
     /// </summary>
     [Order]
-    public SettingsItem<int> CursorRotation { get; private set; }
+    public SettingsItem<float> CursorRotation { get; private set; }
 
     /// <summary>
     /// Toggles a trial for your cursor
@@ -225,7 +225,6 @@ public partial class SettingsProfile
     [Order]
     public SettingsItem<int> FPS { get; private set; }
 
-
     #endregion
 
     #region Audio
@@ -287,11 +286,16 @@ public partial class SettingsProfile
     /// Restarts settings to the game's defaults
     /// </summary>
     public SettingsItem<Variant> ResetToDefaults { get; private set; }
+
     #endregion
+
+    #region Initializers
+
+
 
     public SettingsProfile()
     {
-        #region Initializers
+        #region Gameplay
 
         Sensitivity = new(0.5f)
         {
@@ -442,6 +446,10 @@ public partial class SettingsProfile
                 MaxValue = 120,
             }
         };
+
+        #endregion
+
+        #region Visual
 
         Skin = new("default")
         {
@@ -634,6 +642,10 @@ public partial class SettingsProfile
             }
         };
 
+        #endregion
+
+        #region Video
+
         VideoRenderScale = new(100)
         {
             Id = "VideoRenderScale",
@@ -708,6 +720,10 @@ public partial class SettingsProfile
             UpdateAction = (value, _) => Engine.MaxFps = UnlockFPS ? 0 : FPS
         };
 
+        #endregion
+
+        #region Audio
+
         AutoplayJukebox = new(true)
         {
             Id = "AutoplayJukebox",
@@ -769,6 +785,10 @@ public partial class SettingsProfile
             }
         };
 
+        #endregion
+
+        #region Other
+
         // RhythiaImport = new(default)
         // {
         //     Id = "RhythiaImport",
@@ -821,6 +841,8 @@ public partial class SettingsProfile
 
         updateApproachTime();
     }
+
+    #endregion
 
     /// <summary>
     /// Orders all the <see cref="SettingsItem{T}"/> that is present in the <see cref="SettingsProfile"/>
